ref: main_testファイルを作り、TestMain関数を置くようにした
